requirements.txt
streamlit
google-generativeai
python-dotenv
langdetect
deep-translator
